# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TehTruth Dev Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List registered contracts
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Register a contract with ABI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Contract registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get contract by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Remove a contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /bridges:
    get:
      operationId: listBridgeTransfers
      tags: [bridges]
      summary: List bridge transfers
      responses:
        "200":
          description: List of bridge transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BridgeTransfer"
    post:
      operationId: createBridgeTransfer
      tags: [bridges]
      summary: Log a bridge transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBridgeTransferBody"
      responses:
        "201":
          description: Transfer logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BridgeTransfer"
  /bridges/{id}:
    get:
      operationId: getBridgeTransfer
      tags: [bridges]
      summary: Get bridge transfer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BridgeTransfer"
        "404":
          description: Not found
  /mev:
    get:
      operationId: listMevOperations
      tags: [mev]
      summary: List MEV operations
      responses:
        "200":
          description: List of MEV operations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MevOperation"
    post:
      operationId: createMevOperation
      tags: [mev]
      summary: Create a MEV operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMevOperationBody"
      responses:
        "201":
          description: Operation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MevOperation"
  /mev/{id}:
    get:
      operationId: getMevOperation
      tags: [mev]
      summary: Get MEV operation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Operation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MevOperation"
        "404":
          description: Not found
    patch:
      operationId: updateMevOperation
      tags: [mev]
      summary: Update MEV operation status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMevOperationBody"
      responses:
        "200":
          description: Operation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MevOperation"
        "404":
          description: Not found
...
  /delegations:
    get:
      operationId: listDelegations
      tags: [delegations]
      summary: List EIP-7702 delegations
      responses:
        "200":
          description: List of delegations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Delegation"
    post:
      operationId: createDelegation
      tags: [delegations]
      summary: Register an EIP-7702 delegation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDelegationBody"
      responses:
        "201":
          description: Delegation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Delegation"
  /delegations/{id}:
    get:
      operationId: getDelegation
      tags: [delegations]
      summary: Get delegation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Delegation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Delegation"
        "404":
          description: Not found
    patch:
      operationId: updateDelegation
      tags: [delegations]
      summary: Update delegation (revoke)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDelegationBody"
      responses:
        "200":
          description: Delegation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Delegation"
        "404":
          description: Not found
    delete:
      operationId: deleteDelegation
      tags: [delegations]
      summary: Remove delegation record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /wallets/{id}/modules:
    get:
      operationId: listWalletModules
      tags: [safe-modules]
      summary: List modules enabled on a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SafeModule"
  /safe-modules:
    get:
      operationId: listSafeModules
      tags: [safe-modules]
      summary: List all Safe modules
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SafeModule"
    post:
      operationId: createSafeModule
      tags: [safe-modules]
      summary: Enable a module on a Safe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSafeModuleBody"
      responses:
        "201":
          description: Module enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeModule"
  /safe-modules/{id}:
    get:
      operationId: getSafeModule
      tags: [safe-modules]
      summary: Get module by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Module details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeModule"
        "404":
          description: Not found
    patch:
      operationId: updateSafeModule
      tags: [safe-modules]
      summary: Update module (enable/disable)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSafeModuleBody"
      responses:
        "200":
          description: Module updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeModule"
        "404":
          description: Not found
    delete:
      operationId: deleteSafeModule
      tags: [safe-modules]
      summary: Remove module record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Contract:
      type: object
      properties:
        id:
          type: integer
        address:
          type: string
        name:
          type: string
        chainId:
          type: integer
        abi: {}
        tags:
          type: ["array", "null"]
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - address
        - name
        - chainId
        - abi
        - createdAt
    CreateContractBody:
      type: object
      properties:
        address:
          type: string
        name:
          type: string
        chainId:
          type: integer
        abi: {}
        tags:
          type: ["array", "null"]
          items:
            type: string
      required:
        - address
        - name
        - chainId
        - abi
    BridgeTransfer:
      type: object
      properties:
        id:
          type: integer
        protocol:
          type: string
        sourceChainId:
          type: integer
        destChainId:
          type: integer
        tokenAddress:
          type: string
        tokenSymbol:
          type: string
        amount:
          type: string
        sender:
          type: string
        recipient:
          type: string
        sourceTxHash:
          type: ["string", "null"]
        destTxHash:
          type: ["string", "null"]
        status:
          type: string
        wasPrivate:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - protocol
        - sourceChainId
        - destChainId
        - tokenAddress
        - tokenSymbol
        - amount
        - sender
        - recipient
        - status
        - wasPrivate
        - createdAt
    CreateBridgeTransferBody:
      type: object
      properties:
        protocol:
          type: string
        sourceChainId:
          type: integer
        destChainId:
          type: integer
        tokenAddress:
          type: string
        tokenSymbol:
          type: string
        amount:
          type: string
        sender:
          type: string
        recipient:
          type: string
        sourceTxHash:
          type: ["string", "null"]
        destTxHash:
          type: ["string", "null"]
        status:
          type: string
        wasPrivate:
          type: boolean
      required:
        - protocol
        - sourceChainId
        - destChainId
        - tokenAddress
        - tokenSymbol
        - amount
        - sender
        - recipient
    MevOperation:
      type: object
      properties:
        id:
          type: integer
        opType:
          type: string
        sender:
          type: string
        chainId:
          type: integer
        target:
          type: ["string", "null"]
        calldata:
          type: ["string", "null"]
        value:
          type: ["string", "null"]
        nonce:
          type: ["string", "null"]
        initCode:
          type: ["string", "null"]
        callGasLimit:
          type: ["string", "null"]
        verificationGasLimit:
          type: ["string", "null"]
        preVerificationGas:
          type: ["string", "null"]
        maxFeePerGas:
          type: ["string", "null"]
        maxPriorityFeePerGas:
          type: ["string", "null"]
        paymasterAndData:
          type: ["string", "null"]
        signature:
          type: ["string", "null"]
        paymaster:
          type: ["string", "null"]
        bundler:
          type: ["string", "null"]
        bundlerProvider:
          type: ["string", "null"]
        bundlerUrl:
          type: ["string", "null"]
        entryPointAddress:
          type: ["string", "null"]
        opHash:
          type: ["string", "null"]
        txHash:
          type: ["string", "null"]
        status:
          type: string
        wasPrivate:
          type: boolean
        metadata:
          type: ["object", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - opType
        - sender
        - chainId
        - status
        - wasPrivate
        - createdAt
    CreateMevOperationBody:
      type: object
      properties:
        opType:
          type: string
        sender:
          type: string
        chainId:
          type: integer
        target:
          type: ["string", "null"]
        calldata:
          type: ["string", "null"]
        value:
          type: ["string", "null"]
        nonce:
          type: ["string", "null"]
        initCode:
          type: ["string", "null"]
        callGasLimit:
          type: ["string", "null"]
        verificationGasLimit:
          type: ["string", "null"]
        preVerificationGas:
          type: ["string", "null"]
        maxFeePerGas:
          type: ["string", "null"]
        maxPriorityFeePerGas:
          type: ["string", "null"]
        paymasterAndData:
          type: ["string", "null"]
        signature:
          type: ["string", "null"]
        paymaster:
          type: ["string", "null"]
        bundler:
          type: ["string", "null"]
        bundlerProvider:
          type: ["string", "null"]
        bundlerUrl:
          type: ["string", "null"]
        entryPointAddress:
          type: ["string", "null"]
        status:
          type: string
        wasPrivate:
          type: boolean
        metadata:
          type: ["object", "null"]
      required:
        - opType
        - sender
        - chainId
    UpdateMevOperationBody:
      type: object
      properties:
        status:
          type: string
        opHash:
          type: ["string", "null"]
        txHash:
          type: ["string", "null"]
        paymaster:
          type: ["string", "null"]
        bundler:
          type: ["string", "null"]
        bundlerProvider:
          type: ["string", "null"]
        bundlerUrl:
          type: ["string", "null"]
        entryPointAddress:
          type: ["string", "null"]
...
    Delegation:
      type: object
      properties:
        id:
          type: integer
        eoaAddress:
          type: string
        delegateAddress:
          type: string
        implementationType:
          type: string
        chainId:
          type: integer
        label:
          type: ["string", "null"]
        isActive:
          type: boolean
        txHash:
          type: ["string", "null"]
        wasPrivate:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - eoaAddress
        - delegateAddress
        - implementationType
        - chainId
        - isActive
        - wasPrivate
        - createdAt
    CreateDelegationBody:
      type: object
      properties:
        eoaAddress:
          type: string
        delegateAddress:
          type: string
        implementationType:
          type: string
        chainId:
          type: integer
        label:
          type: ["string", "null"]
        isActive:
          type: boolean
        txHash:
          type: ["string", "null"]
        wasPrivate:
          type: boolean
      required:
        - eoaAddress
        - delegateAddress
        - implementationType
        - chainId
    UpdateDelegationBody:
      type: object
      properties:
        isActive:
          type: boolean
        label:
          type: ["string", "null"]
...
    SafeModule:
      type: object
      properties:
        id:
          type: integer
        safeAddress:
          type: string
        moduleAddress:
          type: string
        chainId:
          type: integer
        moduleType:
          type: string
        label:
          type: ["string", "null"]
        isEnabled:
          type: boolean
        enableTxHash:
          type: ["string", "null"]
        disableTxHash:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - safeAddress
        - moduleAddress
        - chainId
        - moduleType
        - isEnabled
        - createdAt
    CreateSafeModuleBody:
      type: object
      properties:
        safeAddress:
          type: string
        moduleAddress:
          type: string
        chainId:
          type: integer
        moduleType:
          type: string
        label:
          type: ["string", "null"]
        isEnabled:
          type: boolean
        enableTxHash:
          type: ["string", "null"]
      required:
        - safeAddress
        - moduleAddress
        - chainId
        - moduleType
    UpdateSafeModuleBody:
      type: object
      properties:
        isEnabled:
          type: boolean
        label:
          type: ["string", "null"]
        disableTxHash:
          type: ["string", "null"]