# lib/api-zod/src/generated/types/bridgeTransfer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TehTruth Dev Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface BridgeTransfer {
  id: number;
  protocol: string;
  sourceChainId: number;
  destChainId: number;
  tokenAddress: string;
  tokenSymbol: string;
  amount: string;
  sender: string;
  recipient: string;
  /** @nullable */
  sourceTxHash?: string | null;
  /** @nullable */
  destTxHash?: string | null;
  status: string;
  wasPrivate: boolean;
  createdAt: Date;
}