# lib/api-zod/src/generated/types/createContractBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TehTruth Dev Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContractBody {
  address: string;
  name: string;
  chainId: number;
  abi: unknown;
  /** @nullable */
  tags?: string[] | null;
}