# lib/api-zod/src/generated/types/createDelegationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TehTruth Dev Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDelegationBody {
  eoaAddress: string;
  delegateAddress: string;
  implementationType: string;
  chainId: number;
  /** @nullable */
  label?: string | null;
  isActive?: boolean;
  /** @nullable */
  txHash?: string | null;
  wasPrivate?: boolean;
}