# lib/api-zod/src/generated/types/delegation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TehTruth Dev Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Delegation {
  id: number;
  eoaAddress: string;
  delegateAddress: string;
  implementationType: string;
  chainId: number;
  /** @nullable */
  label?: string | null;
  isActive: boolean;
  /** @nullable */
  txHash?: string | null;
  wasPrivate: boolean;
  createdAt: Date;
}